requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-decouple==3.8
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
pyotp==2.9.0
requests==2.32.3
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
prometheus-client==0.21.0
reportlab==4.2.5
pillow==11.0.0

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10  # PostgreSQL adapter
redis==5.2.0

# Celery for background tasks
celery==5.4.0
flower==2.0.1  # Celery monitoring

# Monitoring
sentry-sdk==2.19.2

# Data & ML (for AI features)
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.5.2

# WebSockets
websockets==14.1
